finished version

+++ tools/tuxedoUpload.pl
@@ -15,17 +15,20 @@ my @fileListWithPath;
 my $zipFiles = 0;
 my $tmpDir = '/tmp/repoupload/';
 my $argument;
-my $flavour = '';
+my $flavour;
+my @flavours;
 my $repo = '';
 my $retValue;
 my $errorValue;
+my $testing = 0;
 my %repos = (
-	test => 'testdeb-tuxedo',
+	testdeb => 'testdeb-tuxedo',
 	live => 'deb-tuxedo');
 
 sub usage {
 	print "usage:\n";
-	print "tuxedoUpload.pl [bionic|focal|jammy] [test|live] file1.deb file1.zip\n";
+	print "tuxedoUpload.pl [test] [bionic|focal|jammy] [testdeb|live] [file1.deb|...] [file1.zip|...]\n";
+	print "multiple flavours are possible";
 }
 
 if (! -e $keyfile) {
@@ -47,14 +50,18 @@ if (@ARGV == 0) {
 }
 
 foreach $argument (@ARGV) {
-	if ($argument =~ /(^bionic$|^focal$|^jammy$)/) {
-		$flavour = $argument;
-	} elsif ($argument =~ /(^test$|^live$)/) {
+	if ($argument =~ /(^test$)/) {
+		$testing = 1;
+		print "testmode activated, no changes will be applied to the repos\n";
+	} elsif ($argument =~ /(^bionic$|^focal$|^jammy$)/) {
+		push @flavours, $argument;
+	} elsif ($argument =~ /(^testdeb$|^live$)/) {
 		$repo = $argument;
 	} elsif (($argument =~ /^.*\.deb$/) && (-e $argument)) {
 		print "valid deb-file: $argument found\n";
 		push @fileListWithPath, $argument;
-		push @fileList, basename($argument);
+		$argument = basename($argument);
+		push @fileList, "incoming/$argument";
 	} elsif (($argument =~ /^.*\.zip$/) && (-e $argument)) {
 		print "valid zip-file: $argument found\n";
 		if (! -e $tmpDir) {
@@ -72,13 +79,13 @@ foreach $argument (@ARGV) {
 	}
 }
 
-if ($flavour eq '') {
+if (@flavours == 0) {
 	print "no valid flavour given [bionic|focal|jammy]\n";
 	usage();
 	exit (0);
 }
 if ($repo eq '') {
-	print "no valid repo given [test|live]\n";
+	print "no valid repo given [testdeb|live]\n";
 	usage();
 	exit (0);
 }
@@ -88,7 +95,7 @@ if ($zipFiles) {
 	while (my $file = readdir(DIR)) {
 		next if ($file =~ m/^\./);
 		push (@fileListWithPath, $tmpDir.$file);
-		push (@fileList, $file);
+		push (@fileList, "incoming/$file");
 	}
 	closedir(DIR);
 }
@@ -97,34 +104,41 @@ print "with path: @fileListWithPath\n";
 print "list: @fileList\n";
 
 print "Repo: $repos{$repo}\n";
-print "Flavour: $flavour\n";
+print "Flavours: @flavours\n";
 if (@fileListWithPath == 0) {
 	print "no valid files to transmit given [.deb]\n";
 	usage();
 	exit (0);
 }
 
+# copy files to server
 my $cmd = "scp -i $keyfile @fileListWithPath root\@px02.tuxedo.de:/mnt/repos/$repos{$repo}/ubuntu/incoming/";
-print "command: $cmd\n";
+print "cmd: $cmd\n";
 
-$retValue = `$cmd`;
-$errorValue = ${^CHILD_ERROR_NATIVE};
-rmdir($tmpDir);
-if ($errorValue != 0) {
-	print "some error has happened while uploading\n";
-	print "error value: $errorValue\n";
-	print "return value:\n$retValue";
-	exit (0);
+if (! $testing) {
+	$retValue = `$cmd`;
+	$errorValue = ${^CHILD_ERROR_NATIVE};
+	rmdir($tmpDir);
+	if ($errorValue != 0) {
+		print "some error has happened while uploading\n";
+		print "error value: $errorValue\n";
+		print "return value:\n$retValue";
+		exit (0);
+	}
 }
 
-#TODO implement reprepro
-$cmd = "ssh -i $keyfile root\@px02.tuxedo.de \"cd /mnt/repos/$repos{$repo}/ubuntu/incoming/ && mv @fileList /root/fais-pablo/\"";
-print "cmd: $cmd\n";
-$retValue = `$cmd`;
-$errorValue = ${^CHILD_ERROR_NATIVE};
-if ($errorValue != 0) {
-	print "some error has happened while executing remote command\n";
-	print "error value: $errorValue\n";
-	print "return value:\n$retValue";
-	exit (0);
+# execute reprepro
+foreach $flavour (@flavours) {
+	$cmd = "ssh -i $keyfile root\@px02.tuxedo.de \"cd /mnt/repos/$repos{$repo}/ubuntu/ && reprepro --ask-passphrase -V includedeb $flavour @fileList\"";
+	print "cmd: $cmd\n";
+	if (! $testing) {
+		$retValue = `$cmd`;
+		$errorValue = ${^CHILD_ERROR_NATIVE};
+		if ($errorValue != 0) {
+			print "some error has happened while executing remote command\n";
+			print "error value: $errorValue\n";
+			print "return value:\n$retValue";
+			exit (0);
+		}
+	}
 }
